content/canvas/test/test_2d.voidreturn.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.voidreturn</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(ctx.save() === undefined, "ctx.save() === undefined");
ok(ctx.restore() === undefined, "ctx.restore() === undefined");
ok(ctx.scale(1, 1) === undefined, "ctx.scale(1, 1) === undefined");
ok(ctx.rotate(0) === undefined, "ctx.rotate(0) === undefined");
ok(ctx.translate(0, 0) === undefined, "ctx.translate(0, 0) === undefined");
if (ctx.transform) { // (avoid spurious failures, since the aim here is not to test that all features are supported)
    ok(ctx.transform(1, 0, 0, 1, 0, 0) === undefined, "ctx.transform(1, 0, 0, 1, 0, 0) === undefined");
}
if (ctx.setTransform) {
    ok(ctx.setTransform(1, 0, 0, 1, 0, 0) === undefined, "ctx.setTransform(1, 0, 0, 1, 0, 0) === undefined");
}
ok(ctx.clearRect(0, 0, 0, 0) === undefined, "ctx.clearRect(0, 0, 0, 0) === undefined");
ok(ctx.fillRect(0, 0, 0, 0) === undefined, "ctx.fillRect(0, 0, 0, 0) === undefined");
ok(ctx.strokeRect(0, 0, 0, 0) === undefined, "ctx.strokeRect(0, 0, 0, 0) === undefined");
ok(ctx.beginPath() === undefined, "ctx.beginPath() === undefined");
ok(ctx.closePath() === undefined, "ctx.closePath() === undefined");
ok(ctx.moveTo(0, 0) === undefined, "ctx.moveTo(0, 0) === undefined");
ok(ctx.lineTo(0, 0) === undefined, "ctx.lineTo(0, 0) === undefined");
ok(ctx.quadraticCurveTo(0, 0, 0, 0) === undefined, "ctx.quadraticCurveTo(0, 0, 0, 0) === undefined");
ok(ctx.bezierCurveTo(0, 0, 0, 0, 0, 0) === undefined, "ctx.bezierCurveTo(0, 0, 0, 0, 0, 0) === undefined");
ok(ctx.arcTo(0, 0, 0, 0, 1) === undefined, "ctx.arcTo(0, 0, 0, 0, 1) === undefined");
ok(ctx.rect(0, 0, 0, 0) === undefined, "ctx.rect(0, 0, 0, 0) === undefined");
ok(ctx.arc(0, 0, 1, 0, 0, true) === undefined, "ctx.arc(0, 0, 1, 0, 0, true) === undefined");
ok(ctx.fill() === undefined, "ctx.fill() === undefined");
ok(ctx.stroke() === undefined, "ctx.stroke() === undefined");
ok(ctx.clip() === undefined, "ctx.clip() === undefined");
if (ctx.putImageData) {
    ok(ctx.putImageData(ctx.getImageData(0, 0, 1, 1), 0, 0) === undefined, "ctx.putImageData(ctx.getImageData(0, 0, 1, 1), 0, 0) === undefined");
}
ok(ctx.drawImage(document.getElementById('yellow.png'), 0, 0, 0, 0, 0, 0, 0, 0) === undefined, "ctx.drawImage(document.getElementById('yellow.png'), 0, 0, 0, 0, 0, 0, 0, 0) === undefined");
ok(ctx.drawImage(canvas, 0, 0, 0, 0, 0, 0, 0, 0) === undefined, "ctx.drawImage(canvas, 0, 0, 0, 0, 0, 0, 0, 0) === undefined");
ok(ctx.createLinearGradient(0, 0, 0, 0).addColorStop(0, 'white') === undefined, "ctx.createLinearGradient(0, 0, 0, 0).addColorStop(0, 'white') === undefined");

if (!_deferred) SimpleTest.finish();
});
</script>
<img src="image_yellow.png" id="yellow.png" class="resource">
